+++ jexxa-web/src/main/java/io/jexxa/drivingadapter/rest/openapi/OpenAPIConvention.java
@@ -25,8 +25,8 @@ import java.util.Optional;
 import java.util.Properties;
 import java.util.stream.Stream;
 
-import static io.jexxa.common.JexxaCoreProperties.JEXXA_CONTEXT_NAME;
-import static io.jexxa.common.JexxaCoreProperties.JEXXA_CONTEXT_VERSION;
+import static io.jexxa.properties.JexxaCoreProperties.JEXXA_CONTEXT_NAME;
+import static io.jexxa.properties.JexxaCoreProperties.JEXXA_CONTEXT_VERSION;
 import static io.jexxa.common.facade.logger.SLF4jLogger.getLogger;
 import static io.jexxa.drivingadapter.rest.JexxaWebProperties.JEXXA_REST_OPEN_API_PATH;
 import static org.eclipse.microprofile.openapi.OASFactory.createAPIResponse;